refactor: add domain package and change Car class directory
