+++ src/main/java/racingcar/service/JudgeService.java
@@ -2,7 +2,7 @@ package racingcar.service;
 
 import java.util.ArrayList;
 
-import racingcar.Car;
+import racingcar.domain.Car;
 
 public class JudgeService {
